LayoutTests/inspector/debugger/debug-inlined-scripts-expected.txt

Tests that all inlined scripts from the same document are shown in the same source frame with html script tags. Bug 54544.

Debugger was enabled.
Script source was shown.
Script execution paused.
Call stack:
    0)  (debug-inlined-scripts.html:3)
==Source frame contents start==


  <script>  function f1() { return 0; }; f1();  </script>
==Source frame contents end==
Script execution paused.
Call stack:
    0) f4 (debug-inlined-scripts.html:10)
    1)  (debug-inlined-scripts.html:12)
==Source frame contents start==


  <script>  function f1() { return 0; }; f1();  </script>     <script>function f2() { return 0; }</script><script>
function f3() { return 0; }
</script>

<script>
function f4()
{
    return 0;
}
f4();
</script>
==Source frame contents end==
Page reloaded.
Debugger was disabled.
